move legend button into more menu
