--- frontend/src/lib/components/InsightLegend/InsightLegendButton.scss
@@ -1,10 +1,0 @@
-.InsightLegendButton {
-    display: flex;
-    align-items: center;
-
-    .InsightLegendButton-title {
-        margin-left: 0.5rem;
-        font-size: 14px;
-        font-weight: 500;
-    }
-}
