+++ frontend/src/queries/schema.json
@@ -1277,9 +1277,6 @@
                 "showLastComputationRefresh": {
                     "type": "boolean"
                 },
-                "showLegendButton": {
-                    "type": "boolean"
-                },
                 "showTable": {
                     "type": "boolean"
                 },
