+++ frontend/src/queries/schema.ts
@@ -321,7 +321,6 @@ export interface InsightVizNode extends Node {
     showCorrelationTable?: boolean
     showLastComputation?: boolean
     showLastComputationRefresh?: boolean
-    showLegendButton?: boolean
 }
 
 /** Base class for insight query nodes. Should not be used directly. */
