+++ frontend/src/scenes/experiments/Experiment.tsx
@@ -1019,7 +1019,6 @@ export function Experiment(): JSX.Element {
                                         kind: NodeKind.InsightVizNode,
                                         source: filtersToQueryNode(transformResultFilters(experimentResults.filters)),
                                         showTable: true,
-                                        showLegendButton: false,
                                         showLastComputation: true,
                                         showLastComputationRefresh: false,
                                     }}
